fix esc doesn't close modal if sidebar LHN isn't mounted
